Create a test stage/task
A `make test` task that will execute the pyspelling check that's part of
our pipeline.

Since it's useful to be able to run all stages of our pipeline locally,
the `antora.dockerfile` has been enhanced to support running `pyspelling`.

This is at the cost of additional setup/dependencies on initial building
of the image.

- add `pyspelling` and its dependencies to `anotora.dockerfile`
- define a `test_tasks` list of commands to be run inside container